pyproject.toml
[project]
name = "triton-lrn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.11.0",
    "triton-windows<3.7",
]
